ENH: datasets: Create the datasets subpackage.
Create the new subpackage `datasets`.
Move misc.face, misc.ascent, and misc.electrocardiogram to datasets.
Deprecate misc.face and misc.ascent.
(misc.electrocardiogram has not been in a release yet, so it does
not require deprecation.)
