+++ doc/source/tutorial/signal.rst
@@ -107,7 +107,8 @@ standard-deviation equal to :math:`\sigma_{o}=\left(o+1\right)/12` :
 A function to compute this Gaussian for arbitrary :math:`x` and :math:`o` is
 also available ( :func:`gauss_spline` ). The following code and Figure uses
 spline-filtering to compute an edge-image (the second-derivative of a smoothed
-spline) of a raccoon's face which is an array returned by the command :func:`misc.face`.
+spline) of a raccoon's face which is an array returned by the command
+:func:`datasets.face`.
 The command :func:`sepfir2d` was used to apply a separable two-dimensional FIR
 filter with mirror- symmetric boundary conditions to the spline coefficients.
 This function is ideally suited for reconstructing samples from spline
@@ -118,10 +119,10 @@ conditions.
 .. plot::
 
    >>> import numpy as np
-   >>> from scipy import signal, misc
+   >>> from scipy import signal, datasets
    >>> import matplotlib.pyplot as plt
 
-   >>> image = misc.face(gray=True).astype(np.float32)
+   >>> image = datasets.face(gray=True).astype(np.float32)
    >>> derfilt = np.array([1.0, -2, 1.0], dtype=np.float32)
    >>> ck = signal.cspline2d(image, 8.0)
    >>> deriv = (signal.sepfir2d(ck, derfilt, [1]) +
@@ -310,10 +311,10 @@ enhancing, and edge-detection for an image.
 .. plot::
 
    >>> import numpy as np
-   >>> from scipy import signal, misc
+   >>> from scipy import signal, datasets
    >>> import matplotlib.pyplot as plt
 
-   >>> image = misc.face(gray=True)
+   >>> image = datasets.face(gray=True)
    >>> w = np.zeros((50, 50))
    >>> w[0][0] = 1.0
    >>> w[49][25] = 1.0
@@ -356,10 +357,10 @@ which is often used for blurring.
 .. plot::
 
    >>> import numpy as np
-   >>> from scipy import signal, misc
+   >>> from scipy import signal, datasets
    >>> import matplotlib.pyplot as plt
 
-   >>> image = misc.ascent()
+   >>> image = datasets.ascent()
    >>> w = signal.gaussian(50, 10.0)
    >>> image_new = signal.sepfir2d(image, w, w)
