+++ scipy/datasets/setup.py
@@ -1,0 +1,14 @@
+from __future__ import division, print_function, absolute_import
+
+
+def configuration(parent_package='', top_path=None):
+    from numpy.distutils.misc_util import Configuration
+    config = Configuration('datasets', parent_package, top_path)
+    config.add_data_dir('tests')
+    config.add_data_dir('_data')
+    return config
+
+
+if __name__ == '__main__':
+    from numpy.distutils.core import setup
+    setup(**configuration(top_path='').todict())
