+++ scipy/misc/tests/test_common.py
@@ -4,7 +4,7 @@ import pytest
 from numpy.testing import assert_equal, assert_allclose, assert_almost_equal
 from scipy._lib._numpy_compat import suppress_warnings
 
-from scipy.misc import pade, logsumexp, face, ascent, electrocardiogram
+from scipy.misc import pade, logsumexp, face, ascent
 from scipy.special import logsumexp as sc_logsumexp
 
 
@@ -24,17 +24,12 @@ def test_pade():
 
 
 def test_face():
-    assert_equal(face().shape, (768, 1024, 3))
+    with suppress_warnings() as sup:
+        sup.filter(DeprecationWarning, "`face` is deprecated")
+        assert_equal(face().shape, (768, 1024, 3))
 
 
 def test_ascent():
-    assert_equal(ascent().shape, (512, 512))
-
-
-def test_electrocardiogram():
-    # Test shape, dtype and stats of signal
-    ecg = electrocardiogram()
-    assert ecg.dtype == float
-    assert_equal(ecg.shape, (108000,))
-    assert_almost_equal(ecg.mean(), -0.16510875)
-    assert_almost_equal(ecg.std(), 0.5992473991177294)
+    with suppress_warnings() as sup:
+        sup.filter(DeprecationWarning, "`ascent` is deprecated")
+        assert_equal(ascent().shape, (512, 512))
