+++ scipy/ndimage/fourier.py
@@ -104,12 +104,12 @@ def fourier_gaussian(input, sigma, n=-1, axis=-1, output=None):
 
     Examples
     --------
-    >>> from scipy import ndimage, misc
+    >>> from scipy import ndimage, datasets
     >>> import numpy.fft
     >>> import matplotlib.pyplot as plt
     >>> fig, (ax1, ax2) = plt.subplots(1, 2)
     >>> plt.gray()  # show the filtered result in grayscale
-    >>> ascent = misc.ascent()
+    >>> ascent = datasets.ascent()
     >>> input_ = numpy.fft.fft2(ascent)
     >>> result = ndimage.fourier_gaussian(input_, sigma=4)
     >>> result = numpy.fft.ifft2(result)
@@ -163,12 +163,12 @@ def fourier_uniform(input, size, n=-1, axis=-1, output=None):
 
     Examples
     --------
-    >>> from scipy import ndimage, misc
+    >>> from scipy import ndimage, datasets
     >>> import numpy.fft
     >>> import matplotlib.pyplot as plt
     >>> fig, (ax1, ax2) = plt.subplots(1, 2)
     >>> plt.gray()  # show the filtered result in grayscale
-    >>> ascent = misc.ascent()
+    >>> ascent = datasets.ascent()
     >>> input_ = numpy.fft.fft2(ascent)
     >>> result = ndimage.fourier_uniform(input_, size=20)
     >>> result = numpy.fft.ifft2(result)
@@ -225,12 +225,12 @@ def fourier_ellipsoid(input, size, n=-1, axis=-1, output=None):
 
     Examples
     --------
-    >>> from scipy import ndimage, misc
+    >>> from scipy import ndimage, datasets
     >>> import numpy.fft
     >>> import matplotlib.pyplot as plt
     >>> fig, (ax1, ax2) = plt.subplots(1, 2)
     >>> plt.gray()  # show the filtered result in grayscale
-    >>> ascent = misc.ascent()
+    >>> ascent = datasets.ascent()
     >>> input_ = numpy.fft.fft2(ascent)
     >>> result = ndimage.fourier_ellipsoid(input_, size=20)
     >>> result = numpy.fft.ifft2(result)
@@ -282,12 +282,12 @@ def fourier_shift(input, shift, n=-1, axis=-1, output=None):
 
     Examples
     --------
-    >>> from scipy import ndimage, misc
+    >>> from scipy import ndimage, datasets
     >>> import matplotlib.pyplot as plt
     >>> import numpy.fft
     >>> fig, (ax1, ax2) = plt.subplots(1, 2)
     >>> plt.gray()  # show the filtered result in grayscale
-    >>> ascent = misc.ascent()
+    >>> ascent = datasets.ascent()
     >>> input_ = numpy.fft.fft2(ascent)
     >>> result = ndimage.fourier_shift(input_, shift=200)
     >>> result = numpy.fft.ifft2(result)
