+++ scipy/signal/signaltools.py
@@ -339,8 +339,8 @@ def fftconvolve(in1, in2, mode="full"):
     The `convolve2d` function allows for other types of image boundaries,
     but is far slower.
 
-    >>> from scipy import misc
-    >>> face = misc.face(gray=True)
+    >>> from scipy import datasets
+    >>> face = datasets.face(gray=True)
     >>> kernel = np.outer(signal.gaussian(70, 8), signal.gaussian(70, 8))
     >>> blurred = signal.fftconvolve(face, kernel, mode='same')
 
@@ -997,8 +997,8 @@ def convolve2d(in1, in2, mode='full', boundary='fill', fillvalue=0):
     boundaries.
 
     >>> from scipy import signal
-    >>> from scipy import misc
-    >>> ascent = misc.ascent()
+    >>> from scipy import datasets
+    >>> ascent = datasets.ascent()
     >>> scharr = np.array([[ -3-3j, 0-10j,  +3 -3j],
     ...                    [-10+0j, 0+ 0j, +10 +0j],
     ...                    [ -3+3j, 0+10j,  +3 +3j]]) # Gx + j*Gy
@@ -1084,8 +1084,8 @@ def correlate2d(in1, in2, mode='full', boundary='fill', fillvalue=0):
     image:
 
     >>> from scipy import signal
-    >>> from scipy import misc
-    >>> face = misc.face(gray=True) - misc.face(gray=True).mean()
+    >>> from scipy import datasets
+    >>> face = datasets.face(gray=True) - datasets.face(gray=True).mean()
     >>> template = np.copy(face[300:365, 670:750])  # right eye
     >>> template -= template.mean()
     >>> face = face + np.random.randn(*face.shape) * 50  # add noise
